% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{invert_match}
\alias{invert_match}
\title{Switch location of matches to location of non-matches.}
\usage{
invert_match(loc)
}
\arguments{
\item{loc}{\verb{numeric matrix} of match locations, as from \code{\link[=str_locate_all]{str_locate_all()}}}
}
\value{
Returns a \verb{numeric matrix} giving locations of non-matches.
}
\description{
Invert a matrix of match locations to match the opposite of
what was previously matched.
}
\examples{
numbers <- "1 and 2 and 4 and 456"
and_loc <- str_locate_all(numbers, "and")[[1]]
str_sub(numbers, and_loc[, "start"], and_loc[, "end"])

num_loc <- invert_match(and_loc)
str_sub(numbers, num_loc[, "start"], num_loc[, "end"])
}
