% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{modifiers}
\alias{modifiers}
\alias{perl}
\alias{fixed}
\alias{regex}
\title{Control pattern matching with modifier functions}
\usage{
perl(pattern, ignore_case = FALSE)

fixed(pattern)

regex(pattern, ignore_case = FALSE)
}
\arguments{
\item{pattern}{\code{character}, a pattern to modify the behaviour of.}

\item{ignore_case}{\code{logical}, should case-insensitive matching be used?
Default is \code{FALSE}.}
}
\description{
\describe{
\item{\code{fixed()}}{Match the string literally.}
\item{\code{regex()}}{Match using a POSIX 1003.2 extended regular expression.}
\item{\code{perl()}}{The default, match using a perl-compatible regular expression.}
}
}
