% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{str_replace_na}
\alias{str_replace_na}
\title{Turn NA into "NA"}
\usage{
str_replace_na(string, replacement = "NA")
}
\arguments{
\item{string}{\verb{character vector} of strings.}

\item{replacement}{a single \code{string}.}
}
\description{
Turn NA into "NA"
}
\examples{
str_replace_na(c(NA, "abc", "def"))
}
